WVPlots: Common Plots for Analysis

Select data analysis plots, under a standardized calling interface implemented on top of ggplot2 and plotly. Plots of interest include: ROC, gain curve, scatter plot with marginal distributions, conditioned scatter plot with marginal densities. box and stem with matching theoretical distribution, density with matching theoretical distribution.

Details

For more information:

vignette(package='WVPlots')

RShowDoc('WVPlots_examples',package='WVPlots')

Website: https://github.com/WinVector/WVPlots